90. Concluindo e testando as regras de negocio de autores.

+++ S4_libraryapi/src/main/java/io/github/cursodsouza/libraryapi/service/AutorService.java
@@ -1,7 +1,9 @@
 package io.github.cursodsouza.libraryapi.service;
 
+import io.github.cursodsouza.libraryapi.exceptions.OperacaoNaoPermitida;
 import io.github.cursodsouza.libraryapi.model.Autor;
 import io.github.cursodsouza.libraryapi.repository.AutorRepository;
+import io.github.cursodsouza.libraryapi.repository.LivroRepository;
 import io.github.cursodsouza.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
@@ -13,10 +15,12 @@ import java.util.UUID;
 public class AutorService {
     private final AutorRepository repository;
     private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository, AutorValidator validator){
+    public AutorService(AutorRepository repository, AutorValidator validator, LivroRepository livroRepository){
         this.repository = repository;
         this.validator =  validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
@@ -37,6 +41,9 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if(possueLivro(autor)){
+            throw new OperacaoNaoPermitida("Não é permitido excluir um Autor que possui livro(s) cadastrado(s)!");
+        }
         repository.delete(autor);
     }
 
@@ -54,5 +61,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    private boolean possueLivro(Autor autor){
+        return  livroRepository.existsByAutor(autor);
     }
 }
